pyproject.toml
[project]
name = "extract-image"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "pypdf>=4.3.1",
    "pillow>=10.4.0",
    "cryptography>=44.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/extract_image"]
